+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/core/music/decoders/AmDecoder.java
@@ -78,6 +78,7 @@ public abstract class AmDecoder {
             FILETYPE_TO_DECODER.put("mp3", MP3Decoder::new);
             FILETYPE_TO_DECODER.put("flac", FlacDecoder::new);
             FILETYPE_TO_DECODER.put("ogg", OggDecoder::new);
+            FILETYPE_TO_DECODER.put("wav", WaveDecoder::new);
         }
     }
 }